refactor: many helper functions moved out of main file, drag cursor back to hand png, cursor displays out of window - there is now a canvas state to track where the dragging is done - functions related to altering the cursor are moved into cursor.rs - the hud that shows pixel offset and zoom level is moved to info_hud.rs - a user_action_mode.rs file is made that will contain logic for what mode is the user in (what action are they selecting) - grid background color now set in grid.rs file
